mb/protectli/vault_kbl: merge bootblock versions
