+++ src/mainboard/protectli/vault_kbl/Kconfig
@@ -10,8 +10,7 @@ config BOARD_SPECIFIC_OPTIONS
 	select SEABIOS_ADD_SERCON_PORT_FILE if PAYLOAD_SEABIOS
 	select SOC_INTEL_KABYLAKE
 	select SPI_FLASH_MACRONIX
-	select SUPERIO_ITE_IT8772F if BOARD_PROTECTLI_FW6
-	select SUPERIO_ITE_IT8613E if BOARD_PROTECTLI_FW6D || BOARD_PROTECTLI_FW6E
+	select SUPERIO_ITE_IT8613E
 	select MAINBOARD_HAS_CRB_TPM
 	select HAVE_INTEL_PTT
 	select TPM2
