--- src/mainboard/protectli/vault_kbl/variants/fw6d/Makefile.inc
@@ -1,3 +1,0 @@
-## SPDX-License-Identifier: GPL-2.0-or-later
-
-bootblock-y += bootblock.c
